Time series simplification: Fixes and cleanup
